Revert moving
